# openapi.json
{"openapi":"3.1.0","info":{"title":"Gerchain + NEF RWA Dashboard","version":"1.1.0"},"paths":{"/api/v1/audit/logs":{"get":{"summary":"Get Audit Logs","operationId":"get_audit_logs_api_v1_audit_logs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditLogResponse"}}}}}}},"/api/v1/rwa/status":{"get":{"summary":"Get Rwa Status","operationId":"get_rwa_status_api_v1_rwa_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RWAStatusResponse"}}}}}}},"/api/v1/rwa/register":{"post":{"summary":"Register Rwa Asset","operationId":"register_rwa_asset_api_v1_rwa_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RWAAssetCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Register Rwa Asset Api V1 Rwa Register Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/escrow/status":{"get":{"summary":"Get Escrow Status","operationId":"get_escrow_status_api_v1_escrow_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EscrowStatusResponse"}}}}}}},"/api/v1/escrow/create":{"post":{"summary":"Create Escrow Account","operationId":"create_escrow_account_api_v1_escrow_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EscrowCreateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Create Escrow Account Api V1 Escrow Create Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/escrow/transfer":{"post":{"summary":"Transfer Escrow Funds","operationId":"transfer_escrow_funds_api_v1_escrow_transfer_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EscrowTransferRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Transfer Escrow Funds Api V1 Escrow Transfer Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/dashboard":{"get":{"summary":"Dashboard View","operationId":"dashboard_view_dashboard_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/api/v1/evidence/submit":{"post":{"summary":"Submit Evidence","operationId":"submit_evidence_api_v1_evidence_submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvidenceSubmitRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/evidence/verify-and-release":{"post":{"summary":"Verify And Release","operationId":"verify_and_release_api_v1_evidence_verify_and_release_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvidenceVerifyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AuditLogItem":{"properties":{"id":{"type":"integer","title":"Id"},"action":{"type":"string","title":"Action"},"details":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Details"},"sha256_hash":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sha256 Hash"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"}},"type":"object","required":["id","action","timestamp"],"title":"AuditLogItem"},"AuditLogResponse":{"properties":{"status":{"type":"string","title":"Status"},"count":{"type":"integer","title":"Count"},"data":{"items":{"$ref":"#/components/schemas/AuditLogItem"},"type":"array","title":"Data"}},"type":"object","required":["status","count","data"],"title":"AuditLogResponse"},"EscrowAccountItem":{"properties":{"id":{"type":"integer","title":"Id"},"account_number":{"type":"string","title":"Account Number"},"owner_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Owner Name"},"balance_nef":{"type":"number","title":"Balance Nef"},"status":{"type":"string","title":"Status"}},"type":"object","required":["id","account_number","balance_nef","status"],"title":"EscrowAccountItem"},"EscrowCreateRequest":{"properties":{"account_number":{"type":"string","title":"Account Number"},"owner_name":{"type":"string","title":"Owner Name"},"initial_balance_nef":{"type":"number","title":"Initial Balance Nef"}},"type":"object","required":["account_number","owner_name","initial_balance_nef"],"title":"EscrowCreateRequest"},"EscrowStatusResponse":{"properties":{"status":{"type":"string","title":"Status"},"count":{"type":"integer","title":"Count"},"escrow_accounts":{"items":{"$ref":"#/components/schemas/EscrowAccountItem"},"type":"array","title":"Escrow Accounts"}},"type":"object","required":["status","count","escrow_accounts"],"title":"EscrowStatusResponse"},"EscrowTransferRequest":{"properties":{"from_account":{"type":"string","title":"From Account"},"to_account":{"type":"string","title":"To Account"},"amount_nef":{"type":"number","title":"Amount Nef"},"milestone_ref":{"type":"string","title":"Milestone Ref"}},"type":"object","required":["from_account","to_account","amount_nef","milestone_ref"],"title":"EscrowTransferRequest"},"EvidenceSubmitRequest":{"properties":{"milestone_ref":{"type":"string","title":"Milestone Ref"},"asset_code":{"type":"string","title":"Asset Code"},"herder_account":{"type":"string","title":"Herder Account"},"gps_coordinates":{"type":"string","title":"Gps Coordinates"},"photo_url":{"type":"string","title":"Photo Url"},"commission_act_ref":{"type":"string","title":"Commission Act Ref"}},"type":"object","required":["milestone_ref","asset_code","herder_account","gps_coordinates","photo_url","commission_act_ref"],"title":"EvidenceSubmitRequest"},"EvidenceVerifyRequest":{"properties":{"milestone_ref":{"type":"string","title":"Milestone Ref"},"approved":{"type":"boolean","title":"Approved"},"amount_nef":{"type":"number","title":"Amount Nef"},"from_account":{"type":"string","title":"From Account","default":"NEF-ESCROW-001"},"verifier_notes":{"type":"string","title":"Verifier Notes"}},"type":"object","required":["milestone_ref","approved","amount_nef","verifier_notes"],"title":"EvidenceVerifyRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RWAAssetCreate":{"properties":{"asset_code":{"type":"string","title":"Asset Code"},"asset_type":{"type":"string","title":"Asset Type"},"location":{"type":"string","title":"Location"},"valuation_nef":{"type":"number","title":"Valuation Nef"}},"type":"object","required":["asset_code","asset_type","location","valuation_nef"],"title":"RWAAssetCreate"},"RWAAssetItem":{"properties":{"id":{"type":"integer","title":"Id"},"asset_code":{"type":"string","title":"Asset Code"},"asset_type":{"type":"string","title":"Asset Type"},"location":{"type":"string","title":"Location"},"valuation_nef":{"type":"number","title":"Valuation Nef"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","asset_code","asset_type","location","valuation_nef","created_at"],"title":"RWAAssetItem"},"RWAStatusResponse":{"properties":{"status":{"type":"string","title":"Status"},"count":{"type":"integer","title":"Count"},"rwa_assets":{"items":{"$ref":"#/components/schemas/RWAAssetItem"},"type":"array","title":"Rwa Assets"}},"type":"object","required":["status","count","rwa_assets"],"title":"RWAStatusResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}